+++ src/main/java/com/overops/plugins/sonar/model/EventsJson.java
@@ -1,4 +1,4 @@
-package com.overops.plugins.sonar;
+package com.overops.plugins.sonar.model;
 
 import java.util.List;
 
@@ -10,28 +10,25 @@ public class EventsJson {
 	public void setRule(String rule) {
 		this.rule = rule;
 	}
+
 	public String getRule() {
 		return rule;
 	}
+
 	public void setComponentKey(String componentKey) {
 		this.componentKey = componentKey;
 	}
+
 	public String getComponentKey() {
 		return componentKey;
 	}
+
 	public void setIssues(List<IssueComment> issues) {
 		this.issues = issues;
 	}
+
 	public List<IssueComment> getIssues() {
 		return issues;
 	}
 
 }
-
-/*
-String rule = "overops:event";
-String componentKey = context.project().key() + sourcefile;
-
-ArrayList<IssueComment> issues = new ArrayList<IssueComment>(int size);
-
-*/
